requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
Jinja2>=3.1.0
itsdangerous>=2.2.0
click>=8.1.0
numpy>=1.24
scipy>=1.9
librosa>=0.9.2
matplotlib>=3.6
Pillow>=9.0
soundfile>=0.12
audioread>=3.0
torch
torchvision